+++ src/addons/meta_human_dna/operators.py
@@ -406,12 +406,14 @@ class ConvertSelectedToDna(bpy.types.Operator, MetahumanDnaImportProperties):
         name="Output Folder",
         default="",
         subtype='DIR_PATH',
+        options={'PATH_SUPPORTS_BLEND_RELATIVE'}
     ) # type: ignore
     maps_folder: bpy.props.StringProperty(
         default='',
         name='Maps Folder',
         description='Optionally, this can be set to a folder location for the face wrinkle maps. Textures following the same naming convention as the metahuman source files will be found and set on the materials automatically.',
-        subtype='DIR_PATH'
+        subtype='DIR_PATH',
+        options={'PATH_SUPPORTS_BLEND_RELATIVE'}
     ) # type: ignore
 
     def execute(self, context):
@@ -1394,6 +1396,7 @@ class DuplicateRigLogicInstance(bpy.types.Operator):
         name="New Output Folder", 
         default="",
         subtype='DIR_PATH',
+        options={'PATH_SUPPORTS_BLEND_RELATIVE'}
     ) # type: ignore
 
     def execute(self, context):