+++ src/addons/meta_human_dna/release_notes.md
@@ -6,6 +6,8 @@
 * Fixed Edit bone rotations on the body are not calibrating[#139](https://github.com/poly-hammer/meta-human-dna-addon/issues/139)
 * Fixed Mesh origin validation [#140](https://github.com/poly-hammer/meta-human-dna-addon/issues/140)
 * Fixed Multi-Language support [#141](https://github.com/poly-hammer/meta-human-dna-addon/issues/141)
+* Fixed Seam from Misaligned bones roll on convert [#155](https://github.com/poly-hammer/meta-human-dna-addon/issues/155)
+* Fixed Head LOD to Body LOD mapping. There are twice as many head LODs as body LODs.
 
 
 > [!WARNING]